requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
watchdog==6.0.0
Pillow==11.0.0
qrcode==8.0
pywin32==308
